move homepage to src/pages
