+++ frontend/src/styles/component/homepage.css
